refactor(numbers): 카드 초기화 상태 코드 hook으로 분리
